V1.0 * Added documentation to README.MD files * Renamed home directory to pages with home as a subfolder.
