+++ src/app/footer/footer.component.html
@@ -1,6 +1,6 @@
 <mat-divider class="mat-primary"></mat-divider>
 <div class="mat-caption">
   <p>
-    Site Developed by <a href="http://www.demiurge.us" target="_blank" rel="noopener">Demiurge Software LLC</a>
+    Site Developed by <a href="http://www.mynoc.com" target="_blank" rel="noopener">myNOC LLC</a>
   </p>
 </div>
